# Update to network 3.0 with backward compatibility for network <= 2.8