--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
 dist/
 cabal.sandbox.config
 .cabal-sandbox/
+.stack-work/