--- a/mongoDB.cabal
+++ b/mongoDB.cabal
@@ -40,7 +40,6 @@ Library
                     , conduit-extra
                     , mtl >= 2
                     , cryptohash -any
-                    , network -any
                     , parsec -any
                     , random -any
                     , random-shuffle -any
@@ -65,7 +64,8 @@ Library
      build-depends: network < 2.9
   else
      -- "Network.BSD" has been moved into its own package `network-bsd`
-     build-depends: network-bsd >= 2.7 && < 2.9
+     build-depends: network >= 3.0
+                    , network-bsd >= 2.7 && < 2.9
 
   Exposed-modules:  Database.MongoDB
                     Database.MongoDB.Admin
@@ -119,7 +119,6 @@ Benchmark bench
                     , containers -any
                     , mtl >= 2
                     , cryptohash -any
-                    , network -any
                     , nonce >= 1.0.5
                     , stm
                     , parsec -any
@@ -130,5 +129,14 @@ Benchmark bench
                     , transformers-base >= 0.4.1
                     , hashtables >= 1.1.2.0
                     , criterion
+
+  if flag(_old-network)
+     -- "Network.BSD" is only available in network < 2.9
+     build-depends: network < 2.9
+  else
+     -- "Network.BSD" has been moved into its own package `network-bsd`
+     build-depends: network >= 3.0
+                    , network-bsd >= 2.7 && < 2.9
+
   default-language: Haskell2010
   default-extensions: OverloadedStrings