--- a/stack-ghc82.yaml
+++ b/stack-ghc82.yaml
@@ -1,0 +1,4 @@
+resolver: lts-11.22
+flags:
+  mongoDB:
+    _old-network: true