--- a/stack-ghc86-network3.yaml
+++ b/stack-ghc86-network3.yaml
@@ -1,0 +1,6 @@
+resolver: lts-13.23
+extra-deps:
+- git: git@github.com:hvr/bson.git # https://github.com/mongodb-haskell/bson/pull/18
+  commit: 2fc8d04120c0758201762b8e22254aeb6d574f41
+- network-bsd-2.8.1.0
+- network-3.1.0.0