--- a/stack-ghc86.yaml
+++ b/stack-ghc86.yaml
@@ -1,0 +1,4 @@
+resolver: lts-13.23
+flags:
+  mongoDB:
+    _old-network: true