This PR is more of a request for comments than a PR intended to be merged right away. It uses network 3 and network-bsd (using [this PR](https://github.com/mongodb-haskell/bson/pull/18) on bson). The build succeeds and tests pass with stack. I added a stack.yaml file and made other choices that I expect you would like to do differently before actually merging.

How would you like to move forward in updating this package to network 3? If you would like to keep compatibility with older network versions, would you like to do that via CPP or some other method? Is breaking backward compatibility a possible option?

See Issue #95