@scott-fleischman 2019-05-23
# Benchmark comparison
## master
```
benchmarking insert/1000
time                 93.61 ms   (83.17 ms .. 102.2 ms)
                     0.975 R²   (0.905 R² .. 0.999 R²)
mean                 93.53 ms   (90.32 ms .. 100.2 ms)
std dev              7.229 ms   (3.543 ms .. 11.79 ms)
variance introduced by outliers: 21% (moderately inflated)
```

## PR
```
benchmarking insert/1000
time                 92.81 ms   (87.57 ms .. 106.6 ms)
                     0.972 R²   (0.890 R² .. 1.000 R²)
mean                 94.75 ms   (91.74 ms .. 102.3 ms)
std dev              7.330 ms   (3.163 ms .. 11.47 ms)
variance introduced by outliers: 21% (moderately inflated)
```