@VictorDenisov 2019-05-26
Abandoning network 2.8 will inconvenience us and the community. If we abandon 2.8 and publish a major version bump we will need to support two codebases and port fixes from newer version to another. Stackage nightly builds are still using network 2.8. I would suggest to introduce conditional compilation here. One of breaking changes that we can afford is to remove exports of primitives imported from Network and replace with our own.