@scott-fleischman 2019-05-27
Ok so I understand that you want to keep backward compatibility.

> One of breaking changes that we can afford is to remove exports of primitives imported from Network and replace with our own.

Ok so I will update this PR with an approach that adds a Mongo-specific `PortId` type which internally uses CPP for the different network versions.

Would you like to copy the `connectTo` code from the old Network module too? Or would you prefer to propagate changes that use `Network.Socket` more directly?