@VictorDenisov 2019-05-28
I think we need to hid Network.Socket as well. Otherwise the client of the library will need to do conditional compilation as well.