@scott-fleischman 2019-05-30
Related: https://github.com/commercialhaskell/stackage/issues/4528