@scott-fleischman 2019-05-30
@VictorDenisov let me know if you like the direction in 21cf023854a2bb5d8c81b862fc07d8145ea2ed49 . I still need to fix a few things for older network versions.

Also I'm not sure why the build is failing. It appears to be due to things not related to the code changes. Perhaps the CI script needs updating?