@VictorDenisov 2019-05-30
Yep. The direction looks right to me. I'll take a look at the build scripts this weekend. Looks like mongodb deprecated some of their older versions.