@scott-fleischman 2019-05-30
@VictorDenisov ok the tests pass across the various stack files, and the benchmarks run, all in the same ballpark.