@VictorDenisov 2019-06-12
I suggest to set this flag default to True because stackage build system doesn't know how to set this flag. We need to be compatible with the current stackage by default. If default build starts failing mongodb package will be taken out of nightly builds.