@scott-fleischman 2019-06-12
Yeah best to merge bson PR first. I removed the default setting for the flag following the bson commit.